@@ -3,23 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@tippy.js/react": "^3.1.1",
     "emailjs-com": "^3.2.0",
     "formik": "^2.2.9",
-    "framer-motion": "^7.0.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "framer-motion": "^10.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
     "react-icons": "^4.3.1",
     "react-reveal": "^1.2.2",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "^5.0.0",
+    "react-router-dom": "^6.8.2",
+    "react-scripts": "^5.0.1",
     "react-toastify": "^8.1.0",
     "typewriter-effect": "^2.18.2",
-    "yarn": "^1.22.17",
     "yup": "^0.32.11"
   },
   "scripts": {
