@@ -1,12 +1,13 @@
 import React from "react";
-import { BrowserRouter as Router, Route } from "react-router-dom";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import "../src/components/css/App.css";
-import Projects from "./components/Projects.jsx";
-import Contact from "./components/Contact.jsx";
-import AboutMe from "./components/AboutMe.jsx";
-import Skills from "./components/Skills.jsx";
-import Experience from "./components/Experience.jsx";
-import Navbar from "./components/Navbar.jsx";
+
+const Projects = React.lazy(() => import("./components/Projects.jsx"));
+const AboutMe = React.lazy(() => import("./components/AboutMe.jsx"));
+const Contact = React.lazy(() => import("./components/Contact.jsx"));
+const Skills = React.lazy(() => import("./components/Skills.jsx"));
+const Experience = React.lazy(() => import("./components/Experience.jsx"));
+const Navbar = React.lazy(() => import("./components/Navbar.jsx"));
 
 function App() {
   return (
@@ -14,26 +15,19 @@ function App() {
       <Router>
         <Navbar />
 
-        <Route exact path="/">
-          <AboutMe />
-        </Route>
+        <Routes>
+          <Route exact path="/" element={<AboutMe />}></Route>
 
-        <Route exact path="/projects">
-          <Projects />
-        </Route>
+          <Route exact path="/projects" element={<Projects />}></Route>
 
-        <Route exact path="/contact">
-          <Contact />
-        </Route>
+          <Route exact path="/contact" element={<Contact />}></Route>
 
-        <Route exact path="/skills">
-          <Skills />
-        </Route>
+          <Route exact path="/skills" element={<Skills />}></Route>
 
-        <Route exact path="/experience">
-          <Experience />
-        </Route>
-        <Route path="*" element={<AboutMe />} />
+          <Route exact path="/experience" element={<Experience />}></Route>
+
+          <Route path="*" element={<AboutMe />} />
+        </Routes>
       </Router>
     </div>
   );
