@@ -9,13 +9,13 @@ import {
 import { AiFillMail } from "react-icons/ai";
 import Zoom from "react-reveal/Zoom";
 import { ToastContainer, toast } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
 import { Formik, Form, Field, ErrorMessage } from "formik";
 import * as yup from "yup";
 import Helmet from "react-helmet";
-// import { motion } from "framer-motion";
+import "react-toastify/dist/ReactToastify.css";
+import { motion } from "framer-motion";
 
-const About = () => {
+const Contact = () => {
   const initialValues = {
     fullname: "",
     email: "",
@@ -69,6 +69,62 @@ const About = () => {
     { id: 3, title: "Phone:", text: "91+ 7977368668" },
     { id: 4, title: "Address", text: "Amboli, Andheri(W), Mumbai 400058" },
   ]);
+
+  const socialMedia = [
+    {
+      id: 1,
+      link: "https://www.linkedin.com/in/suraj-yadav-01752b222/",
+      title: "LinkedIn",
+      icon: <FaLinkedinIn className="headerIcon" style={{ color: "#FFF" }} />,
+    },
+    {
+      id: 2,
+      link: "https://github.com/sarry7045",
+      title: "GitHub",
+      icon: <FaGithub className="headerIconGit" style={{ color: "#FFF" }} />,
+    },
+    {
+      id: 3,
+      link: "https://www.instagram.com/suraj.28_11/",
+      title: "Instagram",
+      icon: <FaInstagram className="headerIcon" style={{ color: "#FFF" }} />,
+    },
+    {
+      id: 4,
+      link: "https://www.facebook.com/profile.php?id=100010635822203",
+      title: "Facebook",
+      icon: <FaFacebookF className="headerIcon" style={{ color: "#FFF" }} />,
+    },
+    {
+      id: 5,
+      link: "mailto: sarry7045@gmail.com",
+      title: "Mail",
+      icon: <AiFillMail className="headerIcon" style={{ color: "#FFF" }} />,
+    },
+  ];
+
+  const formikObj = [
+    {
+      id: 1,
+      label: "Full Name:",
+      name: "fullname",
+    },
+    {
+      id: 2,
+      label: "Email Address:",
+      name: "email",
+    },
+    {
+      id: 3,
+      label: "Phone Number:",
+      name: "phone",
+    },
+    {
+      id: 4,
+      label: "Message:",
+      name: "feedback",
+    },
+  ];
   return (
     <>
       <Helmet>
@@ -85,108 +141,50 @@ const About = () => {
       </Helmet>
       <div className="about">
         <div className="container">
-          {/* <motion.div
+          <motion.div
             initial={{ opacity: 0 }}
             animate={{ opacity: 1 }}
-            transition={{ delay: 0.5, duration: 1 }}
+            transition={{ delay: 0.1, duration: 0.9 }}
             className="row  h-650 alignCenter"
-          > */}
-          <Zoom>
-            <div className="row  h-650 alignCenter">
-              <div className="col-6">
-                <div className="about__img">
-                  <img src="/images/Avatar.png" alt="man" loading="lazy" />
-                </div>
+          >
+            <div className="col-6">
+              <div className="about__img">
+                <img src="/Images/Avatar.png" alt="man" loading="lazy" />
               </div>
-              <div className="col-6">
-                <div className="about__info">
-                  {/* <h1>Let's Connect</h1> */}
-                  <h1>Empowering your web identity</h1>
-                  {/* <div className="about__info-p1" style={{ fontSize: "1rem" }}>
-                    "I am an innovative Front-End Developer With Multiple Years
-                    Of experience in Web Development, Software Development and
-                    SEO.
-                  </div> */}
+            </div>
+            <div className="col-6">
+              <div className="about__info">
+                <h1>Empowering your web identity</h1>
 
-                  <div className="info__contacts">
-                    <div className="row">
-                      {state.map((info) => (
-                        <div className="col-6">
-                          <strong>{info.title}</strong>
-                          <p>{info.text}</p>
-                        </div>
-                      ))}
-                      <ul className="contact__header__ul">
-                        <a
-                          href="https://www.linkedin.com/in/suraj-yadav-01752b222/"
-                          target="_blank"
-                          rel="noreferrer"
-                          title="LinkedIn"
-                        >
-                          <li>
-                            <FaLinkedinIn
-                              className="headerIcon"
-                              style={{ color: "#FFF" }}
-                            />
-                          </li>
-                        </a>
-                        <a
-                          href="https://github.com/sarry7045"
-                          target="_blank"
-                          rel="noreferrer"
-                          title="GitHub"
-                        >
-                          <li>
-                            <FaGithub
-                              className="headerIconGit"
-                              style={{ color: "#FFF" }}
-                            />
-                          </li>
-                        </a>
-                        <a
-                          href="https://www.instagram.com/suraj.28_11/"
-                          target="_blank"
-                          rel="noreferrer"
-                          title="Instagram"
-                        >
-                          {" "}
-                          <li>
-                            <FaInstagram
-                              className="headerIcon"
-                              style={{ color: "#FFF" }}
-                            />
-                          </li>
-                        </a>
-                        <a
-                          href="https://www.facebook.com/profile.php?id=100010635822203"
-                          target="_blank"
-                          rel="noreferrer"
-                          title="Facebook"
-                        >
-                          {" "}
-                          <li>
-                            <FaFacebookF
-                              className="headerIcon"
-                              style={{ color: "#FFF" }}
-                            />
-                          </li>
-                        </a>
-                        <a href="mailto: sarry7045@gmail.com" title="Mail">
-                          <li>
-                            <AiFillMail
-                              className="headerIcon"
-                              style={{ color: "#FFF" }}
-                            />
-                          </li>
-                        </a>
-                      </ul>
-                    </div>
+                <div className="info__contacts">
+                  <div className="row">
+                    {state.map((info) => (
+                      <div className="col-6">
+                        <strong>{info.title}</strong>
+                        <p>{info.text}</p>
+                      </div>
+                    ))}
+                    <ul className="contact__header__ul">
+                      {socialMedia.map((item) => {
+                        return (
+                          <>
+                            <a
+                              href={item.link}
+                              target="_blank"
+                              rel="noreferrer"
+                              title={item.title}
+                            >
+                              <li>{item.icon}</li>
+                            </a>
+                          </>
+                        );
+                      })}
+                    </ul>
                   </div>
                 </div>
               </div>
-              {/* </motion.div> */}
             </div>
-          </Zoom>
+          </motion.div>
 
           <div className="wrapper">
             <Zoom>
@@ -198,35 +196,24 @@ const About = () => {
               >
                 <Form onSubmit={sendEmail}>
                   <div className="form">
-                    <div className="inputfield">
-                      <label>Full Name:</label>
-                      <Field type="text" name="fullname" className="input" />
-                      <br />
-                      <p style={{ color: "red" }}>
-                        <ErrorMessage name="fullname" />
-                      </p>
-                    </div>
-                    <div className="inputfield">
-                      <label>Email Address:</label>
-                      <Field type="text" name="email" className="input" />
-                      <p style={{ color: "red" }}>
-                        <ErrorMessage name="email" />
-                      </p>
-                    </div>
-                    <div className="inputfield">
-                      <label>Phone Number:</label>
-                      <Field type="text" name="phone" className="input" />
-                      <p style={{ color: "red" }}>
-                        <ErrorMessage name="phone" />
-                      </p>
-                    </div>
-                    <div className="inputfield">
-                      <label>Message:</label>
-                      <Field type="text" className="textarea" name="feedback" />
-                      <p style={{ color: "red" }}>
-                        <ErrorMessage name="feedback" />
-                      </p>
-                    </div>
+                    {formikObj.map((item) => {
+                      return (
+                        <>
+                          <div className="inputfield">
+                            <label>{item.label}</label>
+                            <Field
+                              type="text"
+                              name={item.name}
+                              className="input"
+                            />
+                            <br />
+                            <p style={{ color: "red" }}>
+                              <ErrorMessage name={item.name} />
+                            </p>
+                          </div>
+                        </>
+                      );
+                    })}
 
                     <div className="inputfield">
                       <input
@@ -248,4 +235,4 @@ const About = () => {
   );
 };
 
-export default About;
+export default Contact;
