@@ -1,7 +1,6 @@
 import React from "react";
 import Helmet from "react-helmet";
-import Zoom from "react-reveal/Zoom";
-// import { motion } from "framer-motion";
+import { motion } from "framer-motion";
 
 const About = () => {
   const [state] = React.useState([
@@ -89,6 +88,7 @@ const About = () => {
       text: "Boston House, Western Express Highyway, Andheri (East) Mumbai 400093",
     },
   ]);
+
   return (
     <>
       <Helmet>
@@ -103,146 +103,101 @@ const About = () => {
       <div className="about">
         <div className="experienceHeight">
           <div className="container">
-            {/* <motion.div
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            transition={{ delay: 0.5, duration: 1 }}
-            className="row  h-650 alignCenter"
-          > */}
-            <div className="row  h-650 alignCenter">
-              <Zoom>
-                {" "}
-                <div
-                  style={{ padding: "1rem 3rem 0 0", margin: "5rem 0 0 0" }}
-                  className="col-6"
-                >
-                  <div className="about__info__Experience">
-                    <h1>IXAR Robotic Solutions</h1>
-                    <div className="commonBorderExp"></div>
-                    <div
-                      className="about__info-p1"
-                      style={{ fontSize: "1rem" }}
-                    >
-                      Worked as a Web Developer and build beautiful
-                      Responsive Website using HTML, CSS, ReactJs, Material-UI etc.
-                    </div>
-                    <div
-                      className="about__info-p2"
-                      style={{ fontSize: "1rem" }}
-                    >
-                      Also Performed SEO
-                      for Website, In order to Rank them on Google Search
-                      Engine.
-                    </div>
+            <motion.div
+              initial={{ opacity: 0 }}
+              animate={{ opacity: 1 }}
+              transition={{ delay: 0.1, duration: 0.9 }}
+              className="row  h-650 alignCenter"
+            >
+              {" "}
+              <div
+                style={{ padding: "1rem 3rem 0 0", margin: "5rem 0 0 0" }}
+                className="col-6"
+              >
+                <div className="about__info__Experience">
+                  <h1>IXAR Robotic Solutions</h1>
+                  <div className="commonBorderExp"></div>
+                  <div className="about__info-p1" style={{ fontSize: "1rem" }}>
+                    Worked as a Web Developer and build beautiful Responsive
+                    Website using HTML, CSS, ReactJs, Material-UI etc.
+                  </div>
+                  <div className="about__info-p2" style={{ fontSize: "1rem" }}>
+                    Also Performed SEO for Website, In order to Rank them on
+                    Google Search Engine.
+                  </div>
 
-                    <div className="info__contacts">
-                      <div className="row">
-                        {state.map((info) => (
-                          <div className="col-6">
-                            <strong>{info.title}</strong>
-                            <p>{info.text}</p>
-                          </div>
-                        ))}
-                      </div>
+                  <div className="info__contacts">
+                    <div className="row">
+                      {state.map((info) => (
+                        <div className="col-6">
+                          <strong>{info.title}</strong>
+                          <p>{info.text}</p>
+                        </div>
+                      ))}
                     </div>
                   </div>
                 </div>
-              </Zoom>
-              <Zoom>
-                <div
-                  style={{ padding: "1rem 3rem 0 0", margin: "5rem 0 0 0" }}
-                  className="col-6"
-                >
-                  <div className="about__info__Experience">
-                    <h1>Payments Gateway Solutions</h1>
-                    <div className="commonBorderExp"></div>
-                    <div
-                      className="about__info-p1"
-                      style={{ fontSize: "1rem" }}
-                    >
-                      Managing UI using HTML, CSS, MaterialUI, JavaScript, and ReactJs.
-                    </div>
-                    <div
-                      className="about__info-p2"
-                      style={{ fontSize: "1rem" }}
-                    >
-                      Worked on Integrating API’s using Redux and all the
-                      Frontend part of M-TRANSACT (Online Payment Software).
-                    </div>{" "}
-                    {/* <div
-                      className="about__info-p2"
-                      style={{ fontSize: "1rem" }}
-                    >
-                      Tools we are using: Redux, Material UI, React-Bootstrap,
-                      Formik, Chart.js, Axios, i18 next, Babel, Webpack, GIT
-                      etc.
-                    </div> */}
-                    <div className="info__contacts">
-                      <div className="row">
-                        {statee.map((infoo) => (
-                          <div className="col-6">
-                            <strong>{infoo.title}</strong>
+              </div>
+              <div
+                style={{ padding: "1rem 3rem 0 0", margin: "5rem 0 0 0" }}
+                className="col-6"
+              >
+                <div className="about__info__Experience">
+                  <h1>Payments Gateway Solutions</h1>
+                  <div className="commonBorderExp"></div>
+                  <div className="about__info-p1" style={{ fontSize: "1rem" }}>
+                    Managing UI using HTML, CSS, MaterialUI, JavaScript, and
+                    ReactJs.
+                  </div>
+                  <div className="about__info-p2" style={{ fontSize: "1rem" }}>
+                    Worked on Integrating API’s using Redux and all the Frontend
+                    part of M-TRANSACT (Online Payment Software).
+                  </div>
+                  <div className="info__contacts">
+                    <div className="row">
+                      {statee.map((infoo) => (
+                        <div className="col-6">
+                          <strong>{infoo.title}</strong>
 
-                            <p>{infoo.text}</p>
-                          </div>
-                        ))}
-                      </div>
+                          <p>{infoo.text}</p>
+                        </div>
+                      ))}
                     </div>
                   </div>
                 </div>
-              </Zoom>
-              <Zoom>
-                <div
-                  style={{ padding: "1rem 3rem 0 0", margin: "5rem 0 0 0" }}
-                  className="col-6"
-                >
-                  <div className="about__info__Experience">
-                    <h1>PhonePe-WealthDesk</h1>
-                    <div className="commonBorderExp"></div>
-                    <div
-                      className="about__info-p1"
-                      style={{ fontSize: "1rem" }}
-                    >
-                      Develop and maintain pages to match business Requirements.
-                    </div>
-                    <div
-                      className="about__info-p2"
-                      style={{ fontSize: "1rem" }}
-                    >
-                      Working on Integrating API’s and all the Frontend part of
-                      WealthBasket's (ETF Portfolios from Experts) and EWG
-                      (Embedded WealthDesk Gateway).
-                    </div>
-                    <div
-                      className="about__info-p2"
-                      style={{ fontSize: "1rem" }}
-                    >
-                      Also Working on Page Optimization.
-                    </div>
-                    {/* <div
-                      className="about__info-p2"
-                      style={{ fontSize: "1rem" }}
-                    >
-                      Tools we are using: Axios, Typescript, Crypto-JS, Material
-                      UI, Sass, Formik, Chart.js, Babel, Webpack, GIT, BitBucket
-                      etc.
-                    </div> */}
-                    <div className="info__contacts">
-                      <div className="row">
-                        {stateee.map((infoo) => (
-                          <div className="col-6">
-                            <strong>{infoo.title}</strong>
+              </div>
+              <div
+                style={{ padding: "1rem 3rem 0 0", margin: "5rem 0 0 0" }}
+                className="col-6"
+              >
+                <div className="about__info__Experience">
+                  <h1>PhonePe-WealthDesk</h1>
+                  <div className="commonBorderExp"></div>
+                  <div className="about__info-p1" style={{ fontSize: "1rem" }}>
+                    Develop and maintain pages to match business Requirements.
+                  </div>
+                  <div className="about__info-p2" style={{ fontSize: "1rem" }}>
+                    Working on Integrating API’s and all the Frontend part of
+                    WealthBasket's (ETF Portfolios from Experts) and EWG
+                    (Embedded WealthDesk Gateway).
+                  </div>
+                  <div className="about__info-p2" style={{ fontSize: "1rem" }}>
+                    Also Working on Page Optimization.
+                  </div>
+                  <div className="info__contacts">
+                    <div className="row">
+                      {stateee.map((infoo) => (
+                        <div className="col-6">
+                          <strong>{infoo.title}</strong>
 
-                            <p>{infoo.text}</p>
-                          </div>
-                        ))}
-                      </div>
+                          <p>{infoo.text}</p>
+                        </div>
+                      ))}
                     </div>
                   </div>
                 </div>
-                {/* </motion.div> */}
-              </Zoom>
-            </div>
+              </div>
+            </motion.div>
           </div>
         </div>
       </div>
